refactor(types) migrate Tally to shared types module
