+++ apps/election-manager/src/components/TallyReportSummary.test.tsx
@@ -1,8 +1,8 @@
 import React from 'react'
 import { render, getByText as domGetByText } from '@testing-library/react'
 
+import { VotingMethod } from '@votingworks/types'
 import TallyReportSummary from './TallyReportSummary'
-import { VotingMethod } from '../config/types'
 
 test('Renders with data source table and voting method table when all data provided', () => {
   const ballotCounts = {
