+++ apps/election-manager/test/helpers/fakeTally.ts
@@ -1,4 +1,4 @@
-import { Tally } from '../../src/config/types'
+import { Tally } from '@votingworks/types'
 
 export default function fakeTally(props: Partial<Tally> = {}): Tally {
   return {
