+++ apps/precinct-scanner/src/api/scan.ts
@@ -1,4 +1,9 @@
-import { AdjudicationReasonInfo, safeParseJSON } from '@votingworks/types'
+import {
+  AdjudicationReasonInfo,
+  BallotSheetInfo,
+  safeParseJSON,
+  CastVoteRecord,
+} from '@votingworks/types'
 import {
   CalibrateResponseSchema,
   GetNextReviewSheetResponse,
@@ -9,7 +14,6 @@ import {
 import { fetchJSON } from '@votingworks/utils'
 import makeDebug from 'debug'
 import {
-  CastVoteRecord,
   RejectedScanningReason,
   ScanningResult,
   ScanningResultType,
