+++ apps/precinct-scanner/src/config/types.ts
@@ -1,4 +1,4 @@
-import { AdjudicationReasonInfo, Dictionary } from '@votingworks/types'
+import { AdjudicationReasonInfo } from '@votingworks/types'
 
 export enum BallotState {
   IDLE = 'idle',
@@ -50,15 +50,4 @@ export interface MachineConfig {
 export interface MachineConfigResponse {
   machineId: string
   codeVersion: string
-}
-
-// Scanner Types
-export interface CastVoteRecord
-  extends Dictionary<string | string[] | boolean> {
-  _precinctId: string
-  _ballotStyleId: string
-  _ballotType: 'absentee' | 'provisional' | 'standard'
-  _ballotId: string
-  _testBallot: boolean
-  _scannerId: string
 }
