+++ libs/types/src/index.ts
@@ -7,3 +7,4 @@ export * from './election'
 export * from './generic'
 export * from './geometry'
 export * from './image'
+export * from './tallies'
