+++ libs/utils/src/votes.test.ts
@@ -37,6 +37,8 @@ test('buildVoteFromCvr', () => {
     _ballotType: 'absentee',
     _ballotStyleId: '1',
     _ballotId: 'b75FfAaktS5jbityDpkFag==',
+    _batchId: '1',
+    _batchLabel: 'Batch 1',
     _testBallot: false,
     _scannerId: 'scanner-6',
     _pageNumber: 1,
@@ -110,6 +112,8 @@ test('buildVoteFromCvr', () => {
     _ballotType: 'absentee',
     _ballotStyleId: '1',
     _ballotId: 'b75FfAaktS5jbityDpkFag==',
+    _batchId: '1',
+    _batchLabel: 'Batch 1',
     _testBallot: false,
     _scannerId: 'scanner-6',
     _pageNumber: 1,
